refactor(studio): improve version menu slightly
