+++ packages/sanity/src/core/studio/components/navbar/resources/ResourcesButton.tsx
@@ -1,12 +1,12 @@
-import {CheckmarkCircleIcon, CopyIcon, HelpCircleIcon} from '@sanity/icons'
-import {Menu, Stack, Text, useToast} from '@sanity/ui'
-import {useCallback, useEffect, useId, useState} from 'react'
+import {HelpCircleIcon} from '@sanity/icons'
+import {Menu} from '@sanity/ui'
+import {useCallback, useState} from 'react'
 import {styled} from 'styled-components'
 
-import {Button, Dialog, MenuButton} from '../../../../../ui-components'
-import {hasSanityPackageInImportMap} from '../../../../environment/hasSanityPackageInImportMap'
-import {Translate, useTranslation} from '../../../../i18n'
+import {Button, MenuButton} from '../../../../../ui-components'
+import {useTranslation} from '../../../../i18n'
 import {SANITY_VERSION} from '../../../../version'
+import {AboutDialog} from './AboutDialog'
 import {useGetHelpResources} from './helper-functions/hooks'
 import {ResourcesMenuItems} from './ResourcesMenuItems'
 
@@ -20,134 +20,22 @@ export function ResourcesButton() {
 
   const {value, error, isLoading} = useGetHelpResources()
   const [aboutDialogOpen, setAboutDialogOpen] = useState(false)
-
-  const latestStudioVersion = value?.latestVersion
   const handleAboutDialogClose = useCallback(() => {
     setAboutDialogOpen(false)
   }, [])
+
   const handleAboutDialogOpen = useCallback(() => {
     setAboutDialogOpen(true)
   }, [])
-  const aboutDialogId = useId()
-  const [copySuccess, setCopySuccess] = useState(false)
-
-  const {push} = useToast()
-  const isAutoUpdating = hasSanityPackageInImportMap()
-
-  const text = `## Current version
-${SANITY_VERSION}
-
-## Latest version
-${latestStudioVersion}
-
-## Auto updates
-${isAutoUpdating ? 'Enabled' : 'Disabled'}
-
-## Page URL
-${document.location.href}
-
-## User agent
-${navigator.userAgent}
-`
-
-  const handleCopyDetails = useCallback(() => {
-    navigator.clipboard.writeText(text).then(
-      () => {
-        setCopySuccess(true)
-      },
-      (err: unknown) => {
-        push({
-          status: 'warning',
-          title: `Unable to write to clipboard: ${(err && typeof err === 'object' && 'message' in err && err.message) || 'unknown error'}`,
-        })
-      },
-    )
-  }, [push, text])
-
-  useEffect(() => {
-    const timer = setTimeout(() => setCopySuccess(false), 3000)
-    return () => clearTimeout(timer)
-  }, [copySuccess])
 
   return (
     <>
       {aboutDialogOpen && (
-        <Dialog
-          header={'About'}
-          width={1}
-          onClickOutside={handleAboutDialogClose}
+        <AboutDialog
+          currentVersion={SANITY_VERSION}
+          latestVersion={value?.latestVersion || 'unknown'}
           onClose={handleAboutDialogClose}
-          id={aboutDialogId}
-        >
-          <Stack space={4}>
-            <Stack space={3}>
-              <Text as="h2" size={1} weight="medium">
-                {t('about-dialog.version-info.current-version.header')}
-              </Text>
-              <Text muted size={1}>
-                {SANITY_VERSION}
-              </Text>
-            </Stack>
-            <Stack space={3}>
-              <Text as="h2" size={1} weight="medium">
-                {t('about-dialog.version-info.current-version.header')}
-              </Text>
-              <Text muted size={1}>
-                <Translate t={t} i18nKey="" components={{}} />
-                {t('about-dialog.version-info.latest-version.text', {latestStudioVersion})}
-                {latestStudioVersion === SANITY_VERSION ? (
-                  <>({t('about-dialog.version-info.up-to-date')})</>
-                ) : (
-                  <>
-                    {' '}
-                    (
-                    <a href="https://www.sanity.io/docs/upgrade">
-                      {t('about-dialog.version-info.how-to-upgrade')}
-                    </a>
-                    )
-                  </>
-                )}
-              </Text>
-            </Stack>
-            <Stack space={3}>
-              <Text as="h2" size={1} weight="medium">
-                {t('about-dialog.version-info.auto-updates.header')}
-              </Text>
-              {isAutoUpdating ? (
-                <Text muted size={1}>
-                  {t('about-dialog.version-info.auto-updates.enabled')}
-                </Text>
-              ) : (
-                <Text muted size={1}>
-                  {t('about-dialog.version-info.auto-updates.disabled')} (
-                  <a href="https://www.sanity.io/docs/auto-updating-studios">
-                    {t('about-dialog.version-info.auto-updates.how-to-enable')}
-                  </a>
-                  )
-                </Text>
-              )}
-            </Stack>
-            <Stack space={3}>
-              <Text as="h2" size={1} weight="medium">
-                {t('about-dialog.version-info.user-agent.header')}
-              </Text>
-              <Text muted size={1}>
-                {navigator.userAgent}
-              </Text>
-            </Stack>
-            <Button
-              icon={copySuccess ? CheckmarkCircleIcon : CopyIcon}
-              mode="bleed"
-              text={
-                copySuccess
-                  ? t('about-dialog.version-info.copy-to-clipboard-button.copied-text')
-                  : t('about-dialog.version-info.copy-to-clipboard-button.text')
-              }
-              paddingY={3}
-              onClick={handleCopyDetails}
-            />
-          </Stack>
-        </Dialog>
+        />
       )}
       <MenuButton
         button={
